добавил функцию получения фото с камеры

+++ handlers/other_handlers.py
@@ -1,6 +1,7 @@
 import asyncio
 import os
 import tempfile
+
 from aiogram import Bot
 from aiogram.dispatcher.router import Router
 from aiogram.filters import CommandStart, Text
@@ -213,30 +214,36 @@ async def alias_list_handler(message: Message):
 
 registered_commands.update(
     [
-        "<code>camera entity_id</code> - получить фото с камеры с id равным entity_id"
+        "<code>camera HA entity_id</code> - получить фото с камеры с идентификатором entity_id"
     ])
 
 
-@router.message(Text(text='camera security_camera'))
+@router.message(lambda message: message.text.lower().startswith("camera ha"))
 async def get_security_camera_image(message: Message, bot: Bot):
     if not await check_access(message):
         return
+    filter_value, value = await Fp.processing_filter(message)
+    if filter_value is not None:
+        req_ha = RequestApi()
+        code, response = req_ha.method_get(f"camera_proxy/camera.{filter_value}", None)
 
-    req_ha = RequestApi()
-    code, response = req_ha.method_get("camera_proxy/camera.security_camera", None)
+        if code == 200:
+            with tempfile.NamedTemporaryFile(delete=False, suffix=".jpg") as tmp_file:
+                tmp_file.write(response.content)
+                tmp_file_path = tmp_file.name
 
-    if code == 200:
-        with tempfile.NamedTemporaryFile(delete=False, suffix=".jpg") as tmp_file:
-            tmp_file.write(response.content)
-            tmp_file_path = tmp_file.name
-
-        try:
-            input_file = FSInputFile(tmp_file_path, filename="camera.jpg")
-            await bot.send_photo(message.from_user.id, photo=input_file)
-        finally:
-            os.remove(tmp_file_path)
+            try:
+                input_file = FSInputFile(tmp_file_path, filename="camera.jpg")
+                await bot.send_photo(message.from_user.id, photo=input_file)
+            finally:
+                os.remove(tmp_file_path)
+        else:
+            await bot.send_message(message.from_user.id,
+                                   f"код ответа: {code}, проверьте работоспособность HA, и корректность запроса")
+            return
     else:
-        await bot.send_message(message.from_user.id, f"Ошибка при получении изображения: {code}")
+        await bot.send_message(message.from_user.id,
+                               f"Запрос отклонен. Пустое значение идентификатора камеры")
 
 
 registered_commands.update(
